Crate open_fastrlp

Re-exports

pub use encode::const_add;

pub use encode::MaxEncodedLen;

pub use encode::MaxEncodedLenAssoc;

Macros

impl_max_encoded_len

Use this to define length of an encoded entity

Structs

Header

Rlp

Enums

DecodeError

Constants

EMPTY_LIST_CODE

EMPTY_STRING_CODE

Traits

BufMut

A trait for values that provide sequential write access to bytes.

Decodable

Encodable

Functions

encode_fixed_size

encode_list

length_of_length

list_length

Derive Macros

RlpDecodable

RlpDecodableWrapper

RlpEncodable

RlpEncodableWrapper

RlpMaxEncodedLen